wildcard ACL: implement ipamHint annotation
we add an annotation on pods to influence the IP they receive.

This is a proof of concept that expects the existing bitmap
to allocate pods in the range:
a:b:c:d:0:0:0:b/112
where b is the offset of the bitmap.

the ipamHint annotation will give an IP
a:b:c:d:xxxx:0:0:b

where xxxx is the ipamHint that is "or"-ed on top of the previous
IP.